graphql-codegen now generating definition files near the query/mutation/fragment definitions
