+++ src/api/index.ts
@@ -1,3 +1,3 @@
 export * from './ApolloProvider';
 export * from './form-error-handling';
-export * from './graphql.generated';
+export * from './schema.generated';
