+++ src/api/schema.generated.ts
@@ -1,10 +1,4 @@
-/* eslint-disable import/no-duplicates, @typescript-eslint/no-empty-interface */
-import * as ApolloReactCommon from '@apollo/client';
-import * as ApolloReactHooks from '@apollo/client';
-import gql from 'graphql-tag';
-
 export type Maybe<T> = T | null;
-
 /** All built-in and custom scalars, mapped to their actual values */
 export interface Scalars {
   ID: string;
@@ -3230,984 +3224,3 @@ export type Zone = Resource &
     name: SecuredString;
     director: SecuredUser;
   };
-
-type DisplayLocation_Country_Fragment = {
-  __typename?: 'Country';
-} & DisplayCountryFragment;
-
-type DisplayLocation_Region_Fragment = {
-  __typename?: 'Region';
-} & DisplayRegionFragment;
-
-type DisplayLocation_Zone_Fragment = {
-  __typename?: 'Zone';
-} & DisplayZoneFragment;
-
-export type DisplayLocationFragment =
-  | DisplayLocation_Country_Fragment
-  | DisplayLocation_Region_Fragment
-  | DisplayLocation_Zone_Fragment;
-
-export type DisplayCountryFragment = { __typename?: 'Country' } & {
-  region: { __typename?: 'SecuredRegion' } & {
-    value?: Maybe<{ __typename?: 'Region' } & DisplayRegionFragment>;
-  };
-} & DisplayPlace_Country_Fragment;
-
-export type DisplayRegionFragment = { __typename?: 'Region' } & {
-  zone: { __typename?: 'SecuredZone' } & {
-    value?: Maybe<{ __typename?: 'Zone' } & DisplayPlace_Zone_Fragment>;
-  };
-} & DisplayPlace_Region_Fragment;
-
-export type DisplayZoneFragment = {
-  __typename?: 'Zone';
-} & DisplayPlace_Zone_Fragment;
-
-type DisplayPlace_Country_Fragment = { __typename?: 'Country' } & {
-  name: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-};
-
-type DisplayPlace_Region_Fragment = { __typename?: 'Region' } & {
-  name: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-};
-
-type DisplayPlace_Zone_Fragment = { __typename?: 'Zone' } & {
-  name: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-};
-
-export type DisplayPlaceFragment =
-  | DisplayPlace_Country_Fragment
-  | DisplayPlace_Region_Fragment
-  | DisplayPlace_Zone_Fragment;
-
-export type LanguageEngagementListItemFragment = {
-  __typename?: 'LanguageEngagement';
-} & Pick<LanguageEngagement, 'id' | 'status'> & {
-    language: { __typename?: 'SecuredLanguage' } & {
-      value?: Maybe<
-        { __typename?: 'Language' } & Pick<Language, 'avatarLetters'> & {
-            name: { __typename?: 'SecuredString' } & Pick<
-              SecuredString,
-              'value'
-            >;
-            displayName: { __typename?: 'SecuredString' } & Pick<
-              SecuredString,
-              'value'
-            >;
-            ethnologuePopulation?: Maybe<
-              { __typename?: 'SecuredInt' } & Pick<SecuredInt, 'value'>
-            >;
-            organizationPopulation?: Maybe<
-              { __typename?: 'SecuredInt' } & Pick<SecuredInt, 'value'>
-            >;
-            rodNumber?: Maybe<
-              { __typename?: 'SecuredInt' } & Pick<SecuredInt, 'value'>
-            >;
-          }
-      >;
-    };
-    products: { __typename?: 'SecuredProductList' } & Pick<
-      SecuredProductList,
-      'total'
-    > & { items: Array<{ __typename?: 'Product' } & Pick<Product, 'type'>> };
-    endDate: { __typename?: 'SecuredDate' } & Pick<SecuredDate, 'value'>;
-    initialEndDate: { __typename?: 'SecuredDate' } & Pick<SecuredDate, 'value'>;
-    completeDate: { __typename?: 'SecuredDate' } & Pick<SecuredDate, 'value'>;
-  };
-
-type ProjectListItem_InternshipProject_Fragment = {
-  __typename?: 'InternshipProject';
-} & Pick<
-  InternshipProject,
-  'id' | 'createdAt' | 'type' | 'sensitivity' | 'status' | 'modifiedAt'
-> & {
-    name: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    deptId: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    step: { __typename?: 'SecuredProjectStep' } & Pick<
-      SecuredProjectStep,
-      'value'
-    >;
-    location: { __typename?: 'SecuredCountry' } & {
-      value?: Maybe<{ __typename?: 'Country' } & DisplayCountryFragment>;
-    };
-    estimatedSubmission: { __typename?: 'SecuredDate' } & Pick<
-      SecuredDate,
-      'value'
-    >;
-  };
-
-type ProjectListItem_TranslationProject_Fragment = {
-  __typename?: 'TranslationProject';
-} & Pick<
-  TranslationProject,
-  'id' | 'createdAt' | 'type' | 'sensitivity' | 'status' | 'modifiedAt'
-> & {
-    name: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    deptId: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    step: { __typename?: 'SecuredProjectStep' } & Pick<
-      SecuredProjectStep,
-      'value'
-    >;
-    location: { __typename?: 'SecuredCountry' } & {
-      value?: Maybe<{ __typename?: 'Country' } & DisplayCountryFragment>;
-    };
-    estimatedSubmission: { __typename?: 'SecuredDate' } & Pick<
-      SecuredDate,
-      'value'
-    >;
-  };
-
-export type ProjectListItemFragment =
-  | ProjectListItem_InternshipProject_Fragment
-  | ProjectListItem_TranslationProject_Fragment;
-
-export interface SessionQueryVariables {}
-
-export type SessionQuery = { __typename?: 'Query' } & {
-  session: { __typename?: 'SessionOutput' } & {
-    user?: Maybe<{ __typename?: 'User' } & LoggedInUserFragment>;
-  };
-};
-
-export type LoggedInUserFragment = { __typename?: 'User' } & Pick<
-  User,
-  'id'
-> & {
-    email: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    timezone: { __typename?: 'SecuredString' } & Pick<SecuredString, 'value'>;
-    realFirstName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-    realLastName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-    displayFirstName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-    displayLastName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-  };
-
-export interface ForgotPasswordMutationVariables {
-  email: Scalars['String'];
-}
-
-export type ForgotPasswordMutation = { __typename?: 'Mutation' } & Pick<
-  Mutation,
-  'forgotPassword'
->;
-
-export interface LoginMutationVariables {
-  input: LoginInput;
-}
-
-export type LoginMutation = { __typename?: 'Mutation' } & {
-  login: { __typename?: 'LoginOutput' } & {
-    user: { __typename?: 'User' } & LoggedInUserFragment;
-  };
-};
-
-export interface LogoutMutationVariables {}
-
-export type LogoutMutation = { __typename?: 'Mutation' } & Pick<
-  Mutation,
-  'logout'
->;
-
-export interface RegisterMutationVariables {
-  input: CreateUserInput;
-}
-
-export type RegisterMutation = { __typename?: 'Mutation' } & {
-  createUser: { __typename?: 'CreateUserOutput' } & {
-    user: { __typename?: 'User' } & LoggedInUserFragment;
-  };
-};
-
-export interface ResetPasswordMutationVariables {
-  input: ResetPasswordInput;
-}
-
-export type ResetPasswordMutation = { __typename?: 'Mutation' } & Pick<
-  Mutation,
-  'resetPassword'
->;
-
-export type UpdateTestUserFragmentFragment = { __typename?: 'User' } & Pick<
-  User,
-  'id'
-> & {
-    realFirstName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-    displayFirstName: { __typename?: 'SecuredString' } & Pick<
-      SecuredString,
-      'value'
-    >;
-  };
-
-export interface GetTestUserQueryVariables {
-  id: Scalars['ID'];
-}
-
-export type GetTestUserQuery = { __typename?: 'Query' } & {
-  user: { __typename?: 'User' } & Pick<User, 'id'> & {
-      realFirstName: { __typename?: 'SecuredString' } & Pick<
-        SecuredString,
-        'value'
-      >;
-    } & UpdateTestUserFragmentFragment;
-};
-
-export interface UpdateTestUserMutationVariables {
-  id: Scalars['ID'];
-  realFirstName: Scalars['String'];
-}
-
-export type UpdateTestUserMutation = { __typename?: 'Mutation' } & {
-  updateUser: { __typename?: 'UpdateUserOutput' } & {
-    user: { __typename?: 'User' } & Pick<User, 'id'> & {
-        realFirstName: { __typename?: 'SecuredString' } & Pick<
-          SecuredString,
-          'value'
-        >;
-        realLastName: { __typename?: 'SecuredString' } & Pick<
-          SecuredString,
-          'value'
-        >;
-      };
-  };
-};
-
-export interface CreateOrganizationMutationVariables {
-  input: CreateOrganizationInput;
-}
-
-export type CreateOrganizationMutation = { __typename?: 'Mutation' } & {
-  createOrganization: { __typename?: 'CreateOrganizationOutput' } & {
-    organization: { __typename?: 'Organization' } & Pick<
-      Organization,
-      'id' | 'createdAt'
-    > & {
-        name: { __typename?: 'SecuredString' } & Pick<
-          SecuredString,
-          'value' | 'canRead' | 'canEdit'
-        >;
-      };
-  };
-};
-
-export interface ProjectsQueryVariables {
-  input: ProjectListInput;
-}
-
-export type ProjectsQuery = { __typename?: 'Query' } & {
-  projects: { __typename?: 'ProjectListOutput' } & Pick<
-    ProjectListOutput,
-    'hasMore' | 'total'
-  > & {
-      items: Array<
-        | ({
-            __typename?: 'InternshipProject';
-          } & ProjectListItem_InternshipProject_Fragment)
-        | ({
-            __typename?: 'TranslationProject';
-          } & ProjectListItem_TranslationProject_Fragment)
-      >;
-    };
-};
-
-export const DisplayPlaceFragmentDoc = gql`
-  fragment DisplayPlace on Place {
-    name {
-      value
-    }
-  }
-`;
-export const DisplayRegionFragmentDoc = gql`
-  fragment DisplayRegion on Region {
-    ...DisplayPlace
-    zone {
-      value {
-        ...DisplayPlace
-      }
-    }
-  }
-  ${DisplayPlaceFragmentDoc}
-`;
-export const DisplayCountryFragmentDoc = gql`
-  fragment DisplayCountry on Country {
-    ...DisplayPlace
-    region {
-      value {
-        ...DisplayRegion
-      }
-    }
-  }
-  ${DisplayPlaceFragmentDoc}
-  ${DisplayRegionFragmentDoc}
-`;
-export const DisplayZoneFragmentDoc = gql`
-  fragment DisplayZone on Zone {
-    ...DisplayPlace
-  }
-  ${DisplayPlaceFragmentDoc}
-`;
-export const DisplayLocationFragmentDoc = gql`
-  fragment DisplayLocation on Location {
-    ... on Country {
-      ...DisplayCountry
-    }
-    ... on Region {
-      ...DisplayRegion
-    }
-    ... on Zone {
-      ...DisplayZone
-    }
-  }
-  ${DisplayCountryFragmentDoc}
-  ${DisplayRegionFragmentDoc}
-  ${DisplayZoneFragmentDoc}
-`;
-export const LanguageEngagementListItemFragmentDoc = gql`
-  fragment LanguageEngagementListItem on LanguageEngagement {
-    id
-    status
-    language {
-      value {
-        name {
-          value
-        }
-        displayName {
-          value
-        }
-        ethnologuePopulation {
-          value
-        }
-        organizationPopulation {
-          value
-        }
-        rodNumber {
-          value
-        }
-        avatarLetters
-      }
-    }
-    products {
-      total
-      items {
-        type
-      }
-    }
-    endDate {
-      value
-    }
-    initialEndDate {
-      value
-    }
-    completeDate {
-      value
-    }
-  }
-`;
-export const ProjectListItemFragmentDoc = gql`
-  fragment ProjectListItem on Project {
-    id
-    createdAt
-    name {
-      value
-    }
-    type
-    sensitivity
-    deptId {
-      value
-    }
-    step {
-      value
-    }
-    status
-    location {
-      value {
-        ...DisplayCountry
-      }
-    }
-    estimatedSubmission {
-      value
-    }
-    modifiedAt
-  }
-  ${DisplayCountryFragmentDoc}
-`;
-export const LoggedInUserFragmentDoc = gql`
-  fragment LoggedInUser on User {
-    id
-    email {
-      value
-    }
-    timezone {
-      value
-    }
-    realFirstName {
-      value
-    }
-    realLastName {
-      value
-    }
-    displayFirstName {
-      value
-    }
-    displayLastName {
-      value
-    }
-  }
-`;
-export const UpdateTestUserFragmentFragmentDoc = gql`
-  fragment UpdateTestUserFragment on User {
-    id
-    realFirstName {
-      value
-    }
-    displayFirstName {
-      value
-    }
-  }
-`;
-export const SessionDocument = gql`
-  query Session {
-    session(browser: true) {
-      user {
-        ...LoggedInUser
-      }
-    }
-  }
-  ${LoggedInUserFragmentDoc}
-`;
-
-/**
- * __useSessionQuery__
- *
- * To run a query within a React component, call `useSessionQuery` and pass it any options that fit your needs.
- * When your component renders, `useSessionQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useSessionQuery({
- *   variables: {
- *   },
- * });
- */
-export function useSessionQuery(
-  baseOptions?: ApolloReactHooks.QueryHookOptions<
-    SessionQuery,
-    SessionQueryVariables
-  >
-) {
-  return ApolloReactHooks.useQuery<SessionQuery, SessionQueryVariables>(
-    SessionDocument,
-    baseOptions
-  );
-}
-export function useSessionLazyQuery(
-  baseOptions?: ApolloReactHooks.LazyQueryHookOptions<
-    SessionQuery,
-    SessionQueryVariables
-  >
-) {
-  return ApolloReactHooks.useLazyQuery<SessionQuery, SessionQueryVariables>(
-    SessionDocument,
-    baseOptions
-  );
-}
-export type SessionQueryHookResult = ReturnType<typeof useSessionQuery>;
-export type SessionLazyQueryHookResult = ReturnType<typeof useSessionLazyQuery>;
-export type SessionQueryResult = ApolloReactCommon.QueryResult<
-  SessionQuery,
-  SessionQueryVariables
->;
-export const ForgotPasswordDocument = gql`
-  mutation ForgotPassword($email: String!) {
-    forgotPassword(email: $email)
-  }
-`;
-export type ForgotPasswordMutationFn = ApolloReactCommon.MutationFunction<
-  ForgotPasswordMutation,
-  ForgotPasswordMutationVariables
->;
-
-/**
- * __useForgotPasswordMutation__
- *
- * To run a mutation, you first call `useForgotPasswordMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useForgotPasswordMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [forgotPasswordMutation, { data, loading, error }] = useForgotPasswordMutation({
- *   variables: {
- *      email: // value for 'email'
- *   },
- * });
- */
-export function useForgotPasswordMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    ForgotPasswordMutation,
-    ForgotPasswordMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<
-    ForgotPasswordMutation,
-    ForgotPasswordMutationVariables
-  >(ForgotPasswordDocument, baseOptions);
-}
-export type ForgotPasswordMutationHookResult = ReturnType<
-  typeof useForgotPasswordMutation
->;
-export type ForgotPasswordMutationResult = ApolloReactCommon.MutationResult<
-  ForgotPasswordMutation
->;
-export type ForgotPasswordMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  ForgotPasswordMutation,
-  ForgotPasswordMutationVariables
->;
-export const LoginDocument = gql`
-  mutation Login($input: LoginInput!) {
-    login(input: $input) {
-      user {
-        ...LoggedInUser
-      }
-    }
-  }
-  ${LoggedInUserFragmentDoc}
-`;
-export type LoginMutationFn = ApolloReactCommon.MutationFunction<
-  LoginMutation,
-  LoginMutationVariables
->;
-
-/**
- * __useLoginMutation__
- *
- * To run a mutation, you first call `useLoginMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLoginMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [loginMutation, { data, loading, error }] = useLoginMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useLoginMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    LoginMutation,
-    LoginMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<LoginMutation, LoginMutationVariables>(
-    LoginDocument,
-    baseOptions
-  );
-}
-export type LoginMutationHookResult = ReturnType<typeof useLoginMutation>;
-export type LoginMutationResult = ApolloReactCommon.MutationResult<
-  LoginMutation
->;
-export type LoginMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  LoginMutation,
-  LoginMutationVariables
->;
-export const LogoutDocument = gql`
-  mutation Logout {
-    logout
-  }
-`;
-export type LogoutMutationFn = ApolloReactCommon.MutationFunction<
-  LogoutMutation,
-  LogoutMutationVariables
->;
-
-/**
- * __useLogoutMutation__
- *
- * To run a mutation, you first call `useLogoutMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useLogoutMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [logoutMutation, { data, loading, error }] = useLogoutMutation({
- *   variables: {
- *   },
- * });
- */
-export function useLogoutMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    LogoutMutation,
-    LogoutMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<LogoutMutation, LogoutMutationVariables>(
-    LogoutDocument,
-    baseOptions
-  );
-}
-export type LogoutMutationHookResult = ReturnType<typeof useLogoutMutation>;
-export type LogoutMutationResult = ApolloReactCommon.MutationResult<
-  LogoutMutation
->;
-export type LogoutMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  LogoutMutation,
-  LogoutMutationVariables
->;
-export const RegisterDocument = gql`
-  mutation Register($input: CreateUserInput!) {
-    createUser(input: $input) {
-      user {
-        ...LoggedInUser
-      }
-    }
-  }
-  ${LoggedInUserFragmentDoc}
-`;
-export type RegisterMutationFn = ApolloReactCommon.MutationFunction<
-  RegisterMutation,
-  RegisterMutationVariables
->;
-
-/**
- * __useRegisterMutation__
- *
- * To run a mutation, you first call `useRegisterMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useRegisterMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [registerMutation, { data, loading, error }] = useRegisterMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useRegisterMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    RegisterMutation,
-    RegisterMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<
-    RegisterMutation,
-    RegisterMutationVariables
-  >(RegisterDocument, baseOptions);
-}
-export type RegisterMutationHookResult = ReturnType<typeof useRegisterMutation>;
-export type RegisterMutationResult = ApolloReactCommon.MutationResult<
-  RegisterMutation
->;
-export type RegisterMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  RegisterMutation,
-  RegisterMutationVariables
->;
-export const ResetPasswordDocument = gql`
-  mutation ResetPassword($input: ResetPasswordInput!) {
-    resetPassword(input: $input)
-  }
-`;
-export type ResetPasswordMutationFn = ApolloReactCommon.MutationFunction<
-  ResetPasswordMutation,
-  ResetPasswordMutationVariables
->;
-
-/**
- * __useResetPasswordMutation__
- *
- * To run a mutation, you first call `useResetPasswordMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useResetPasswordMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [resetPasswordMutation, { data, loading, error }] = useResetPasswordMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useResetPasswordMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    ResetPasswordMutation,
-    ResetPasswordMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<
-    ResetPasswordMutation,
-    ResetPasswordMutationVariables
-  >(ResetPasswordDocument, baseOptions);
-}
-export type ResetPasswordMutationHookResult = ReturnType<
-  typeof useResetPasswordMutation
->;
-export type ResetPasswordMutationResult = ApolloReactCommon.MutationResult<
-  ResetPasswordMutation
->;
-export type ResetPasswordMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  ResetPasswordMutation,
-  ResetPasswordMutationVariables
->;
-export const GetTestUserDocument = gql`
-  query GetTestUser($id: ID!) {
-    user(id: $id) {
-      id
-      realFirstName {
-        value
-      }
-      ...UpdateTestUserFragment
-    }
-  }
-  ${UpdateTestUserFragmentFragmentDoc}
-`;
-
-/**
- * __useGetTestUserQuery__
- *
- * To run a query within a React component, call `useGetTestUserQuery` and pass it any options that fit your needs.
- * When your component renders, `useGetTestUserQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useGetTestUserQuery({
- *   variables: {
- *      id: // value for 'id'
- *   },
- * });
- */
-export function useGetTestUserQuery(
-  baseOptions?: ApolloReactHooks.QueryHookOptions<
-    GetTestUserQuery,
-    GetTestUserQueryVariables
-  >
-) {
-  return ApolloReactHooks.useQuery<GetTestUserQuery, GetTestUserQueryVariables>(
-    GetTestUserDocument,
-    baseOptions
-  );
-}
-export function useGetTestUserLazyQuery(
-  baseOptions?: ApolloReactHooks.LazyQueryHookOptions<
-    GetTestUserQuery,
-    GetTestUserQueryVariables
-  >
-) {
-  return ApolloReactHooks.useLazyQuery<
-    GetTestUserQuery,
-    GetTestUserQueryVariables
-  >(GetTestUserDocument, baseOptions);
-}
-export type GetTestUserQueryHookResult = ReturnType<typeof useGetTestUserQuery>;
-export type GetTestUserLazyQueryHookResult = ReturnType<
-  typeof useGetTestUserLazyQuery
->;
-export type GetTestUserQueryResult = ApolloReactCommon.QueryResult<
-  GetTestUserQuery,
-  GetTestUserQueryVariables
->;
-export const UpdateTestUserDocument = gql`
-  mutation UpdateTestUser($id: ID!, $realFirstName: String!) {
-    updateUser(input: { user: { id: $id, realFirstName: $realFirstName } }) {
-      user {
-        id
-        realFirstName {
-          value
-        }
-        realLastName {
-          value
-        }
-      }
-    }
-  }
-`;
-export type UpdateTestUserMutationFn = ApolloReactCommon.MutationFunction<
-  UpdateTestUserMutation,
-  UpdateTestUserMutationVariables
->;
-
-/**
- * __useUpdateTestUserMutation__
- *
- * To run a mutation, you first call `useUpdateTestUserMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useUpdateTestUserMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [updateTestUserMutation, { data, loading, error }] = useUpdateTestUserMutation({
- *   variables: {
- *      id: // value for 'id'
- *      realFirstName: // value for 'realFirstName'
- *   },
- * });
- */
-export function useUpdateTestUserMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    UpdateTestUserMutation,
-    UpdateTestUserMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<
-    UpdateTestUserMutation,
-    UpdateTestUserMutationVariables
-  >(UpdateTestUserDocument, baseOptions);
-}
-export type UpdateTestUserMutationHookResult = ReturnType<
-  typeof useUpdateTestUserMutation
->;
-export type UpdateTestUserMutationResult = ApolloReactCommon.MutationResult<
-  UpdateTestUserMutation
->;
-export type UpdateTestUserMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  UpdateTestUserMutation,
-  UpdateTestUserMutationVariables
->;
-export const CreateOrganizationDocument = gql`
-  mutation createOrganization($input: CreateOrganizationInput!) {
-    createOrganization(input: $input) {
-      organization {
-        id
-        name {
-          value
-          canRead
-          canEdit
-        }
-        createdAt
-      }
-    }
-  }
-`;
-export type CreateOrganizationMutationFn = ApolloReactCommon.MutationFunction<
-  CreateOrganizationMutation,
-  CreateOrganizationMutationVariables
->;
-
-/**
- * __useCreateOrganizationMutation__
- *
- * To run a mutation, you first call `useCreateOrganizationMutation` within a React component and pass it any options that fit your needs.
- * When your component renders, `useCreateOrganizationMutation` returns a tuple that includes:
- * - A mutate function that you can call at any time to execute the mutation
- * - An object with fields that represent the current status of the mutation's execution
- *
- * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
- *
- * @example
- * const [createOrganizationMutation, { data, loading, error }] = useCreateOrganizationMutation({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useCreateOrganizationMutation(
-  baseOptions?: ApolloReactHooks.MutationHookOptions<
-    CreateOrganizationMutation,
-    CreateOrganizationMutationVariables
-  >
-) {
-  return ApolloReactHooks.useMutation<
-    CreateOrganizationMutation,
-    CreateOrganizationMutationVariables
-  >(CreateOrganizationDocument, baseOptions);
-}
-export type CreateOrganizationMutationHookResult = ReturnType<
-  typeof useCreateOrganizationMutation
->;
-export type CreateOrganizationMutationResult = ApolloReactCommon.MutationResult<
-  CreateOrganizationMutation
->;
-export type CreateOrganizationMutationOptions = ApolloReactCommon.BaseMutationOptions<
-  CreateOrganizationMutation,
-  CreateOrganizationMutationVariables
->;
-export const ProjectsDocument = gql`
-  query Projects($input: ProjectListInput!) {
-    projects(input: $input) {
-      items {
-        ...ProjectListItem
-      }
-      hasMore
-      total
-    }
-  }
-  ${ProjectListItemFragmentDoc}
-`;
-
-/**
- * __useProjectsQuery__
- *
- * To run a query within a React component, call `useProjectsQuery` and pass it any options that fit your needs.
- * When your component renders, `useProjectsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useProjectsQuery({
- *   variables: {
- *      input: // value for 'input'
- *   },
- * });
- */
-export function useProjectsQuery(
-  baseOptions?: ApolloReactHooks.QueryHookOptions<
-    ProjectsQuery,
-    ProjectsQueryVariables
-  >
-) {
-  return ApolloReactHooks.useQuery<ProjectsQuery, ProjectsQueryVariables>(
-    ProjectsDocument,
-    baseOptions
-  );
-}
-export function useProjectsLazyQuery(
-  baseOptions?: ApolloReactHooks.LazyQueryHookOptions<
-    ProjectsQuery,
-    ProjectsQueryVariables
-  >
-) {
-  return ApolloReactHooks.useLazyQuery<ProjectsQuery, ProjectsQueryVariables>(
-    ProjectsDocument,
-    baseOptions
-  );
-}
-export type ProjectsQueryHookResult = ReturnType<typeof useProjectsQuery>;
-export type ProjectsLazyQueryHookResult = ReturnType<
-  typeof useProjectsLazyQuery
->;
-export type ProjectsQueryResult = ApolloReactCommon.QueryResult<
-  ProjectsQuery,
-  ProjectsQueryVariables
->;
