+++ src/scenes/Authentication/Login/Login.tsx
@@ -2,8 +2,9 @@ import { FORM_ERROR } from 'final-form';
 import React, { useEffect, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 import { Except } from 'type-fest';
-import { handleFormError, useLoginMutation } from '../../../api';
+import { handleFormError } from '../../../api';
 import { useSession } from '../../../components/Session';
+import { useLoginMutation } from './Login.generated';
 import { LoginForm, LoginFormProps as Props } from './LoginForm';
 
 export const Login = (props: Except<Props, 'onSubmit'>) => {
