+++ src/scenes/Authentication/Logout/Logout.tsx
@@ -2,8 +2,8 @@ import { CircularProgress } from '@material-ui/core';
 import * as React from 'react';
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useLogoutMutation } from '../../../api';
 import { useSession } from '../../../components/Session';
+import { useLogoutMutation } from './logout.generated';
 
 export const Logout = () => {
   const navigate = useNavigate();
