--- src/scenes/DevTest/fragment.graphql
@@ -1,9 +1,0 @@
-fragment UpdateTestUserFragment on User {
-  id
-  realFirstName {
-    value
-  }
-  displayFirstName {
-    value
-  }
-}
