--- src/scenes/DevTest/queries.graphql
@@ -1,23 +1,0 @@
-query GetTestUser($id: ID!) {
-  user(id: $id) {
-    id
-    realFirstName {
-      value
-    }
-    ...UpdateTestUserFragment
-  }
-}
-
-mutation UpdateTestUser($id: ID!, $realFirstName: String!) {
-  updateUser(input: { user: { id: $id, realFirstName: $realFirstName } }) {
-    user {
-      id
-      realFirstName {
-        value
-      }
-      realLastName {
-        value
-      }
-    }
-  }
-}
